implemented delete link

+++ models/link.go
@@ -21,7 +21,6 @@ func GetLinksByUser(id string) (*[]Link, error) {
 func (link *Link) Put(uid string, field string, value interface{}) error {
 	db.First(link)
 	if link.OwnerID != uid {
-		log.Println("owner: ", link.OwnerID, "user: ", uid)
 		return fmt.Errorf("User doesn't own resource")
 	}
 	log.Println("Updating link")
@@ -50,6 +49,10 @@ func (link *Link) Update() error {
 }
 
 // Delete a db entry2
-func (link *Link) Delete() error {
+func (link *Link) Delete(uid string) error {
+	db.First(link)
+	if link.OwnerID != uid {
+		return fmt.Errorf("User doesn't own resource")
+	}
 	return db.Delete(link).Error
 }
